fhir/lung-cancer/requirements.txt
absl-py==0.7.1
tensorflow==1.13.1
google-cloud-storage==1.16.0
google-api-python-client==1.7.8
oauth2client==4.1.3
